% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortgagePayment.R
\docType{methods}
\name{mortgagePayment}
\alias{mortgagePayment}
\title{Calculate payment for mortgage for a fixed rate.}
\usage{
mortgagePayment(
  loan,
  amortPeriod,
  interestRate,
  payFreq,
  term = NA,
  prepayPct = NA,
  prepayMount = NA
)
}
\arguments{
\item{loan}{numeric, Loan amount}

\item{amortPeriod}{numeric, Number of months for the amortization period.}

\item{interestRate}{numeric, The annual interest rate (\%).}

\item{payFreq}{character, Can be one of \code{monthly}, \code{semi-monthly},
\code{weekly}, \code{biweekly}. Currently, only support \code{monthly}.}

\item{term}{numeric, Number of years for the current term. If missing, the term is
the same as the amortPeriod.}

\item{prepayPct}{numeric, Prepayment percentage, it supports multiple options for comparison.
If missing, this is no prepare plan.}

\item{prepayMount}{numeric, Prepayment amount, it supports multiple options for comparison.
If missing, this is no prepare plan.}
}
\value{
each payment
}
\description{
This function is to calculate mortgage payment for a fixed rate.
}
\note{
This function produce similar but different results when comparing to
mortgage calculators online (see descriptions).
}
\examples{
\dontrun{
## loan 364000 for 20 years (240 months) with an interest rate of 0.359 no term specification
paymenttable <- mortgagePayment(loan = 364000,
                            amortPeriod = 240,
                            interestRate = 4.6,
                            payFreq = "monthly")

## loan 364000 for 20 years (240 months) with an interest rate of 0.359 term = 5years
paymenttable <- mortgagePayment(loan = 364000,
                            amortPeriod = 240,
                            interestRate = 4.6,
                            payFreq = "monthly",
                            term = 5)
## loan 364000 for 20 years (240 months) with an interest rate of 0.359 term = 5years
## with prepay options
paymenttable <- mortgagePayment(loan = 364000,
                            amortPeriod = 240,
                            interestRate = 4.79,
                            payFreq = "monthly",
                            term = 5,
                            prepayMount = c(100, 120, 150, 170, 210))


}


}
\author{
Yong Luo
}
